[ :wrench: | #64 | fix ] Product 파트 추가 및 수정
- ProductRepository 에 saveAll() 추가
- Product 엔티티의 Integer -> Long 으로 수정

+++ src/main/java/com/objects/marketbridge/domain/model/Product.java
@@ -24,16 +24,16 @@ public class Product extends BaseEntity{
 
     private String name;
 
-    private Integer price;
+    private Long price;
 
     private boolean isSubs;
 
     private String thumbImg;
 
-    private Integer discountRate;
+    private Long discountRate;
 
     @Builder
-    private Product(Category categoryId, boolean isOwn, String name, Integer price, boolean isSubs, String thumbImg, Integer discountRate) {
+    private Product(Category categoryId, boolean isOwn, String name, Long price, boolean isSubs, String thumbImg, Long discountRate) {
         this.categoryId = categoryId;
         this.isOwn = isOwn;
         this.name = name;
